% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat0.R
\name{cat0}
\alias{cat0}
\title{Concatenate and Print without Spaces}
\usage{
cat0(..., sep = "")
}
\arguments{
\item{...}{\R object(s), and any additional arguments to \code{\link[=cat]{cat()}}.}

\item{sep}{Strings to append after each object.}
}
\value{
None (invisible NULL).
#export
}
\description{
Concatenate and Print without Spaces
}
\examples{
#cat(letters)
#cat("\n")
#cat0(letters)


}
