% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprint.R
\name{iprint}
\alias{iprint}
\title{Printing function for RMD chunks.}
\usage{
iprint(..., newline = TRUE)
}
\arguments{
\item{...}{R object(s)}

\item{newline}{logical. Whether or not to go to a
new line at the end of printing; defaults to \code{TRUE}.}
}
\value{
Nothing (actually an invisible \code{NULL}; see
\code{\link[=cat]{cat()}} for details).
}
\description{
\code{iprint()} prints object(s) using \code{\link[=cat]{cat()}}, without spaces
in-between objects, and with an optional return at the end.
Ideal for labeling output in a chunk when working in RMD.
}
\details{
This function is not worth using over \code{\link[=cat]{cat()}} if you are
printing to the console (even if you don't want spaces).
}
\examples{
## default (newline=TRUE)
iprint(c("a","b","c"))
iprint(c("x","y","z"))
## no return at the end
iprint(c("a","b","c"), newline=FALSE)
iprint(c("x","y","z"), newline=FALSE)
}
