% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izzy_html.R
\name{izzy_html}
\alias{izzy_html}
\title{HTML Rendering for package \code{isabelletsmith}}
\usage{
izzy_html(default_name = "default", toc = TRUE, toc_depth = 4, ...)
}
\arguments{
\item{default_name}{string. Name of HTML template, if not "default".}

\item{toc}{logical. Whether or not to include table of contents.}

\item{toc_depth}{integer. Number of heading levels to include in table of contents.}

\item{...}{other arguments to be passed to \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}.}
}
\value{
R Markdown output format, to use with \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
\code{izzy_html()} prepares an RMD document to be rendered into HTML.
It gets CSS and HTML template files from the \code{isabelletsmith} package
(\url{https://github.com/isabelle-smith/isabelletsmith}), then passes
those and any additional args into \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}.
}
\examples{
## base case
# rmarkdown::render("file.Rmd", izzy_html())

## changing TOC defaults
# rmarkdown::render("file.Rmd", izzy_html(toc=FALSE, toc_depth=5))

## different template name
# rmarkdown::render("file.Rmd", izzy_html(default_name="007-default"))

## additional argumentss
# rmarkdown::render("file.Rmd", izzy_html(code_folding="hide"))
}
